chore: move lnclient models
