+++ lsp/lsp_service.go
@@ -13,7 +13,7 @@ import (
 	"strings"
 	"time"
 
-	"github.com/getAlby/nostr-wallet-connect/models/lnclient"
+	"github.com/getAlby/nostr-wallet-connect/lnclient"
 	"github.com/getAlby/nostr-wallet-connect/service"
 	"github.com/sirupsen/logrus"
 )
